refactor(trigger): extract webhook serving into webhook.go
Pure move of the webhook trigger surface out of engine.go: path
registration, kind: Task and kind: PipelineTask HTTP dispatch, HMAC
signature + replay verification, and the leaked HTTP/UI helpers (task
error page, SDK/SPA injection, static asset serving). No behavior
change; functions, types, and consts relocated verbatim.

Part of #388 (god-file decomposition).
